# containers/reasoning-engine/openapi.yaml
openapi: 3.0.3
info:
  title: Reasoning Engine API
  version: 1.0.0
  description: |
    Advanced reasoning capabilities service.
    
    ## Authentication
    All endpoints require JWT authentication via Bearer token.
    
    ## Reasoning Types
    - Chain-of-Thought: Sequential reasoning with steps
    - Tree-of-Thought: Multi-branch exploration
    - Analogical: Pattern matching and adaptation
    - Counterfactual: Scenario analysis
    - Causal: Cause-effect analysis
    - Probabilistic: Uncertainty quantification
    - Meta-Reasoning: Reasoning about reasoning

servers:
  - url: /api/v1
    description: API Version 1

tags:
  - name: Reasoning
    description: Reasoning task management
  - name: Health
    description: Health check endpoints

paths:
  /reasoning/reason:
    post:
      summary: Synchronous reason (inline)
      description: |
        Run reasoning and return steps and conclusion without creating a persisted task (dataflow §10.1).
        Use for inline explainability (e.g. risk, recommendations).
      tags: [Reasoning]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  description: Question or problem to reason about
                context:
                  type: array
                  items:
                    type: string
                type:
                  type: string
                  enum: [chain_of_thought, tree_of_thought, analogical, counterfactual, causal, probabilistic, meta_reasoning, custom]
      responses:
        '200':
          description: Reasoning result
          content:
            application/json:
              schema:
                type: object
                properties:
                  steps:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        order: { type: number }
                        type: { type: string }
                        content: { type: string }
                        reasoning: { type: string }
                        confidence: { type: number }
                  reasoning:
                    type: string
                  conclusion:
                    type: string
                  confidence:
                    type: number
                  alternatives:
                    type: array
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /reasoning/tasks:
    post:
      summary: Create reasoning task
      description: Create a new reasoning task
      tags: [Reasoning]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, input]
              properties:
                name:
                  type: string
                description:
                  type: string
                type:
                  type: string
                  enum: [chain_of_thought, tree_of_thought, analogical, counterfactual, causal, probabilistic, meta_reasoning, custom]
                input:
                  type: object
                  required: [query]
                  properties:
                    query:
                      type: string
                    context:
                      type: array
                      items:
                        type: string
                    constraints:
                      type: object
                    examples:
                      type: array
                      items:
                        type: string
                options:
                  type: object
                  properties:
                    maxDepth:
                      type: number
                    maxBranches:
                      type: number
                    includeAlternatives:
                      type: boolean
                    minConfidence:
                      type: number
                      minimum: 0
                      maximum: 1
      responses:
        '201':
          description: Reasoning task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasoningTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    get:
      summary: List reasoning tasks
      description: List reasoning tasks with filtering
      tags: [Reasoning]
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, reasoning, completed, failed, cancelled]
        - name: limit
          in: query
          schema:
            type: number
            default: 100
        - name: continuationToken
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of reasoning tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReasoningTask'
                  continuationToken:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  
  /reasoning/tasks/{id}:
    get:
      summary: Get reasoning task
      description: Get reasoning task by ID
      tags: [Reasoning]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reasoning task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasoningTask'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    put:
      summary: Update reasoning task
      description: Update reasoning task
      tags: [Reasoning]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [pending, reasoning, completed, failed, cancelled]
                output:
                  type: object
                metadata:
                  type: object
                error:
                  type: string
      responses:
        '200':
          description: Reasoning task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasoningTask'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  
  /reasoning/tasks/{id}/cancel:
    post:
      summary: Cancel reasoning task
      description: Cancel a reasoning task
      tags: [Reasoning]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reasoning task cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasoningTask'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  
  /health:
    get:
      summary: Health check
      description: Liveness probe
      tags: [Health]
      responses:
        '200':
          description: Service is healthy
  
  /ready:
    get:
      summary: Readiness check
      description: Readiness probe
      tags: [Health]
      responses:
        '200':
          description: Service is ready
        '503':
          description: Service is not ready

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  schemas:
    ReasoningTask:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
        status:
          type: string
        input:
          type: object
        output:
          type: object
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
  
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
    
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
    
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object